+++ daemon/src/config.rs
@@ -17,7 +17,6 @@ pub struct Config {
     pub gfx_last_mode: GfxVendors,
     pub gfx_managed: bool,
     pub gfx_vfio_enable: bool,
-    pub gfx_save_compute_vfio: bool,
     pub active_profile: String,
     pub toggle_profiles: Vec<String>,
     #[serde(skip)]
@@ -38,7 +37,6 @@ impl Default for Config {
             gfx_last_mode: GfxVendors::Hybrid,
             gfx_managed: true,
             gfx_vfio_enable: false,
-            gfx_save_compute_vfio: true,
             active_profile: "normal".into(),
             toggle_profiles: vec!["normal".into(), "boost".into(), "silent".into()],
             curr_fan_mode: 0,
@@ -69,6 +67,11 @@ impl Config {
             } else {
                 if let Ok(data) = serde_json::from_str(&buf) {
                     return data;
+                } else if let Ok(data) = serde_json::from_str::<ConfigV352>(&buf) {
+                    let config = data.into_current();
+                    config.write();
+                    info!("Updated config version to: {}", VERSION);
+                    return config;
                 } else if let Ok(data) = serde_json::from_str::<ConfigV341>(&buf) {
                     let config = data.into_current();
                     config.write();