+++ daemon/src/config_old.rs
@@ -28,7 +28,6 @@ impl ConfigV317 {
             gfx_last_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
             gfx_vfio_enable: false,
-            gfx_save_compute_vfio: false,
             active_profile: self.active_profile,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.curr_fan_mode,
@@ -57,7 +56,6 @@ impl ConfigV324 {
             gfx_last_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
             gfx_vfio_enable: false,
-            gfx_save_compute_vfio: false,
             active_profile: self.active_profile,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.curr_fan_mode,
@@ -87,7 +85,38 @@ impl ConfigV341 {
             gfx_last_mode: GfxVendors::Hybrid,
             gfx_managed: self.gfx_managed,
             gfx_vfio_enable: false,
-            gfx_save_compute_vfio: false,
+            active_profile: self.active_profile,
+            toggle_profiles: self.toggle_profiles,
+            curr_fan_mode: self.curr_fan_mode,
+            bat_charge_limit: self.bat_charge_limit,
+            power_profiles: self.power_profiles,
+        }
+    }
+}
+
+
+#[derive(Deserialize, Serialize)]
+pub struct ConfigV352 {
+    pub gfx_mode: GfxVendors,
+    pub gfx_last_mode: GfxVendors,
+    pub gfx_managed: bool,
+    pub gfx_vfio_enable: bool,
+    pub gfx_save_compute_vfio: bool,
+    pub active_profile: String,
+    pub toggle_profiles: Vec<String>,
+    #[serde(skip)]
+    pub curr_fan_mode: u8,
+    pub bat_charge_limit: u8,
+    pub power_profiles: BTreeMap<String, Profile>,
+}
+
+impl ConfigV352 {
+    pub(crate) fn into_current(self) -> Config {
+        Config {
+            gfx_mode: GfxVendors::Hybrid,
+            gfx_last_mode: GfxVendors::Hybrid,
+            gfx_managed: self.gfx_managed,
+            gfx_vfio_enable: false,
             active_profile: self.active_profile,
             toggle_profiles: self.toggle_profiles,
             curr_fan_mode: self.curr_fan_mode,